Add files via upload
This CS50 PSet0 project, developed in Scratch, served as an engaging introduction to fundamental programming concepts. While visually simple, creating this interactive game was a valuable exercise in understanding logical sequencing, event handling, and how basic computational instructions are processed. It was a fun and effective way to grasp the core principles of how a program flows. Feel free to import the .sb3 file into scratch.mit.edu to explore!